ArrayField: use a stable key for items [+]
This is a bit of a tricky one. Issue inveniosoftware/invenio-app-rdm#1146
summarizes the observed problem. The source of it was in bad interaction
between TextField, formik and list of React components. tldr; we didn't use
stable keys.

In order to keep the benefits of TextField's optimizations, I fixed the
key problem (see https://reactjs.org/docs/lists-and-keys.html#keys) by
generating artificial keys in react-invenio-deposit's deserialization
of concerned fields and by making any new item in the array
be injected with a key that will never cross those artificially generated
ones in react-invenio-deposit. This ends up giving us stable keys which in
turn makes our React list behave correctly.