overridable: add dynamicParametrize function
* Added a function to 'dynamically' parametrize a component, based on
values from the context. Currently, this includes the field values
stored in Formik, but we can easily add more values from the React
context as needed.

* The user can implement a function that receives these input values and
returns the override props for a given `Overridable` component. They
need to import the `dynamicParametrize` function from the top level of
this module, as well as the component they want to override.

* `dynamicParametrize` might currently throw an error if used on a
component that's rendered outside of the Formik context